add exceptions PacienteService

+++ src/main/java/com/digitalhouse/proyectofinal/service/impl/PacienteService.java
@@ -1,6 +1,8 @@
 package com.digitalhouse.proyectofinal.service.impl;
 
 import com.digitalhouse.proyectofinal.entity.Paciente;
+import com.digitalhouse.proyectofinal.exception.BadRequestException;
+import com.digitalhouse.proyectofinal.exception.ResourceNotFoundException;
 import com.digitalhouse.proyectofinal.repository.IPacienteRepository;
 import com.digitalhouse.proyectofinal.service.IPacienteService;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,9 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public Paciente guardarPaciente(Paciente paciente) {
+        if(paciente == null || paciente.getNombre() == null || paciente.getApellido() ==null){
+            throw new BadRequestException("El paciente no puede ser nulo y debe tener un nombre y apellido.");
+        }
         return pacienteRepository.save(paciente);
     }
 
@@ -33,11 +38,17 @@ public class PacienteService implements IPacienteService {
 
     @Override
     public void modificarPaciente(Paciente paciente) {
+        if(paciente == null || paciente.getId()==null){
+            throw new BadRequestException("El paciente no puede ser nulo y debe tener un ID.");
+        }
         pacienteRepository.save(paciente);
     }
 
     @Override
     public void eliminarPaciente(Integer id) {
+        if(!pacienteRepository.existsById(id)){
+            throw new ResourceNotFoundException("No se encontró el paciente con el ID " + id);
+        }
         pacienteRepository.deleteById(id);
     }
